chore: split CLI into multiple files
